// archive/wallet/core/src/types.ts
export interface SignedTransaction {
  bodyBytes: Uint8Array;
  authInfoBytes: Uint8Array;
  signatures: Uint8Array[];
}
// ...
export interface GasEstimation {
  gasLimit: string;
  feeAmount: string;
  feeDenom: string;
}
// ...
export interface MsgCreatePool {
  creator: string;
  tokenA: string;
  tokenB: string;
  amountA: string;
  amountB: string;
}

export interface MsgAddLiquidity {
  provider: string;
  poolId: number;
  amountA: string;
  amountB: string;
}

export interface MsgRemoveLiquidity {
  provider: string;
  poolId: number;
  shares: string;
}

export interface MsgSwap {
  trader: string;
  poolId: number;
  tokenIn: string;
  tokenOut: string;
  amountIn: string;
  minAmountOut: string;
}

// Oracle Message Types
export interface MsgSubmitPrice {
  validator: string;
  feeder: string;
  asset: string;
  price: string;
}

export interface MsgDelegateFeedConsent {
  validator: string;
  delegate: string;
}

// Standard Cosmos Messages
export interface MsgSend {
  fromAddress: string;
  toAddress: string;
  amount: Array<{ denom: string; amount: string }>;
}

export interface MsgDelegate {
  delegatorAddress: string;
  validatorAddress: string;
  amount: { denom: string; amount: string };
}

export interface MsgUndelegate {
  delegatorAddress: string;
  validatorAddress: string;
  amount: { denom: string; amount: string };
}

export interface MsgBeginRedelegate {
  delegatorAddress: string;
  validatorSrcAddress: string;
  validatorDstAddress: string;
  amount: { denom: string; amount: string };
}

export interface MsgWithdrawDelegatorReward {
  delegatorAddress: string;
  validatorAddress: string;
}

export interface MsgVote {
  proposalId: string;
  voter: string;
  option: number;
}